% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_masslist.R
\name{make_pfas_db}
\alias{make_pfas_db}
\title{Make a PFAS structure database from a list of input files}
\usage{
make_pfas_db(
  mollist_path = NULL,
  biotrans_jarfile = NULL,
  result_path = NULL,
  keep_sdf = FALSE
)
}
\arguments{
\item{mollist_path}{A path to a directory containing .xls molecule lists for
import}

\item{biotrans_jarfile}{A path to a BioTransformer jar file}

\item{result_path}{A directory to store results (defaults to currect working
directory)}

\item{keep_sdf}{T/F should intermediate sdf files be preserved}
}
\value{
Filenames were results are written
}
\description{
Reads input files, performs transformations, writes results to database
(optionally keeps intermediate sdf files)
}
