% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_masslist.R
\name{make_pfas_flat}
\alias{make_pfas_flat}
\title{Make a flat file from the PFAScreeneR.db file}
\usage{
make_pfas_flat(
  db_file,
  tables = c("molecules", "list_membership", "mol_props")
)
}
\arguments{
\item{db_file}{Full path to a database file}

\item{tables}{List of table names to merge and write to file}
}
\value{
filename of the exported flat file
}
\description{
Useful for creating flat files to use as input to other software
(e.g., To make a mass list in Thermo Compound Discoverer)
}
